requirements.txt
kopf
smolagents[toolkit,litellm,mcp]
pyyaml
requests
python-dotenv
kubernetes
